Search Post with sequelize operator

+++ src/migrations/create-post.js
@@ -49,10 +49,10 @@ module.exports = {
         type: Sequelize.STRING,
       },
       priceNumber: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.FLOAT,
       },
       areaNumber: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.FLOAT,
       },
       createdAt: {
         allowNull: false,
